Starting from the begining again - sidepane and startmenu are combined and their state is shared
