+++ src/reducers/index.js
@@ -1,5 +1,6 @@
 import { combineReducers, createStore } from "redux";
 
+import combinedReducer from "../actions/acts";
 import appReducer from "./apps";
 import deskReducer from "./desktop";
 import fileReducer from "./files";
@@ -7,8 +8,6 @@ import folderAndFileReducer from "./folderAndFile";
 import globalReducer from "./globals";
 import menusReducer from "./menu";
 import settReducer from "./settings";
-import paneReducer from "./sidepane";
-import menuReducer from "./startmenu";
 import taskReducer from "./taskbar";
 import wallReducer from "./wallpaper";
 import widReducer from "./widpane";
@@ -17,8 +16,7 @@ const allReducers = combineReducers({
   wallpaper: wallReducer,
   taskbar: taskReducer,
   desktop: deskReducer,
-  startmenu: menuReducer,
-  sidepane: paneReducer,
+  combined: combinedReducer,
   widpane: widReducer,
   apps: appReducer,
   menus: menusReducer,
