+++ src/reducers/sidepane.js
@@ -49,24 +49,53 @@ const defState = {
 };
 
 const paneReducer = (state = defState, action) => {
-  if (action.type == "PANETHEM") {
-    var tmpState = { ...state };
-    tmpState.quicks[4].src = action.payload;
-    return tmpState;
-  } else if (action.type == "BANDTOGG") {
-    return { ...state, banhide: !state.banhide };
-  } else if (action.type == "BANDHIDE") {
-    return { ...state, banhide: true };
-  } else if (action.type == "PANETOGG") {
-    return { ...state, hide: !state.hide };
-  } else if (action.type == "PANEHIDE") {
-    return { ...state, hide: true };
-  } else if (action.type == "CALNTOGG") {
-    return { ...state, calhide: !state.calhide };
-  } else if (action.type == "CALNHIDE") {
-    return { ...state, calhide: true };
-  } else {
-    return state;
+  switch (action.type) {
+    case "PANETHEM":
+      return {
+        ...state,
+        quicks: state.quicks.map((item, index) =>
+          index === 4 ? { ...item, src: action.payload } : item
+        ),
+      };
+
+    case "BANDTOGG":
+      return {
+        ...state,
+        banhide: !state.banhide,
+      };
+
+    case "BANDHIDE":
+      return {
+        ...state,
+        banhide: true,
+      };
+
+    case "PANETOGG":
+      return {
+        ...state,
+        hide: !state.hide,
+      };
+
+    case "PANEHIDE":
+      return {
+        ...state,
+        hide: true,
+      };
+
+    case "CALNTOGG":
+      return {
+        ...state,
+        calhide: !state.calhide,
+      };
+
+    case "CALNHIDE":
+      return {
+        ...state,
+        calhide: true,
+      };
+
+    default:
+      return state;
   }
 };
